+++ libr/anal/arch/8051/8051_ass.c
@@ -1,5 +1,4 @@
 /* radare - LGPL - Copyright 2009-2019 - hmht */
-#include"8051_ass.h"
 
 /*****************************************************************************\
  *              Architecture
