+++ libr/anal/p/8051.mk
@@ -1,5 +1,4 @@
 OBJ_8051=anal_8051.o
-CFLAGS+=-I../asm/arch/8051
 
 STATIC_OBJ+=${OBJ_8051}
 #OBJ_8051+=../../../../../../../../../../../../../../../../../../../../${LTOP}/asm/arch/8051/8051.o
